6.12 move steps to the corresponding package
