+++ android/app/build.gradle
@@ -96,4 +96,7 @@ dependencies {
 
     //Settings
     implementation 'androidx.preference:preference:1.1.1'
+
+    // Recycler view
+    implementation 'com.android.support:recyclerview-v7:28.0.0'
 }
