+++ android/app/src/main/java/com/github/braillesystems/learnbraille/ui/screens/theory/TheoryFreeNavigation.kt
@@ -11,7 +11,8 @@ import com.github.braillesystems.learnbraille.data.entities.*
 import com.github.braillesystems.learnbraille.data.repository.MutableTheoryRepository
 import com.github.braillesystems.learnbraille.data.repository.TheoryRepository
 import com.github.braillesystems.learnbraille.ui.screens.menu.MenuFragmentDirections
-import com.github.braillesystems.learnbraille.ui.screens.theory.AbstractStepFragment.Companion.stepArgName
+import com.github.braillesystems.learnbraille.ui.screens.theory.steps.AbstractStepFragment
+import com.github.braillesystems.learnbraille.ui.screens.theory.steps.AbstractStepFragment.Companion.stepArgName
 import com.github.braillesystems.learnbraille.utils.*
 import kotlinx.coroutines.launch
 import org.koin.android.ext.android.get
