+++ android/app/src/main/java/com/github/braillesystems/learnbraille/ui/screens/theory/steps/InputSymbolFragment.kt
@@ -1,27 +1,30 @@
-package com.github.braillesystems.learnbraille.ui.screens.theory
+package com.github.braillesystems.learnbraille.ui.screens.theory.steps
 
 import android.os.Bundle
 import android.os.Vibrator
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.core.content.getSystemService
-import androidx.core.text.parseAsHtml
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.ViewModelProvider
 import com.github.braillesystems.learnbraille.R
 import com.github.braillesystems.learnbraille.data.entities.BrailleDots
-import com.github.braillesystems.learnbraille.data.entities.InputDots
-import com.github.braillesystems.learnbraille.data.entities.spelling
+import com.github.braillesystems.learnbraille.data.entities.Input
+import com.github.braillesystems.learnbraille.data.entities.Symbol
 import com.github.braillesystems.learnbraille.data.repository.PreferenceRepository
-import com.github.braillesystems.learnbraille.databinding.FragmentLessonsInputDotsBinding
+import com.github.braillesystems.learnbraille.databinding.FragmentLessonsInputSymbolBinding
 import com.github.braillesystems.learnbraille.ui.screens.*
+import com.github.braillesystems.learnbraille.ui.screens.theory.getStepArg
+import com.github.braillesystems.learnbraille.ui.screens.theory.toCurrentStep
+import com.github.braillesystems.learnbraille.ui.screens.theory.toNextStep
+import com.github.braillesystems.learnbraille.ui.screens.theory.toPrevStep
 import com.github.braillesystems.learnbraille.ui.views.*
 import com.github.braillesystems.learnbraille.utils.application
 import com.github.braillesystems.learnbraille.utils.checkedBuzz
 import org.koin.android.ext.android.inject
 import timber.log.Timber
 
-class InputDotsFragment : AbstractStepFragment(R.string.lessons_help_input_dots) {
+class InputSymbolFragment : AbstractStepFragment(R.string.lessons_help_input_symbol) {
 
     private lateinit var expectedDots: BrailleDots
     private lateinit var dotsState: BrailleDotsState
@@ -34,26 +37,27 @@ class InputDotsFragment : AbstractStepFragment(R.string.lessons_help_input_dots)
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ) = DataBindingUtil.inflate<FragmentLessonsInputDotsBinding>(
+    ) = DataBindingUtil.inflate<FragmentLessonsInputSymbolBinding>(
         inflater,
-        R.layout.fragment_lessons_input_dots,
+        R.layout.fragment_lessons_input_symbol,
         container,
         false
     ).apply {
 
-        Timber.i("Start initialize input dots fragment")
+        Timber.i("Initialize input symbol fragment")
 
         val step = getStepArg()
-        require(step.data is InputDots)
-        infoTextView.text = step.data.text?.parseAsHtml()
-            ?: getString(R.string.lessons_input_dots_info_template)
-                .format(step.data.dots.spelling)
-        brailleDots.dotsState.display(step.data.dots)
+        require(step.data is Input)
+        require(step.data.material.data is Symbol)
+        val symbol = step.data.material.data
+        letter.text = symbol.symbol.toString()
+        brailleDots.dotsState.display(symbol.brailleDots)
+        makeIntroLetterToast(symbol.symbol.toString())
 
-        updateStepTitle(step.lessonId, step.id, R.string.lessons_title_input_dots)
+        updateStepTitle(step.lessonId, step.id, R.string.lessons_title_input_symbol)
         setHasOptionsMenu(true)
 
-        expectedDots = step.data.dots
+        expectedDots = symbol.brailleDots
         userTouchedDots = false
         dotsState = brailleDots.dotsState.apply {
             uncheck()
@@ -66,17 +70,21 @@ class InputDotsFragment : AbstractStepFragment(R.string.lessons_help_input_dots)
         }
 
 
-        val viewModelFactory = InputViewModelFactory(application, expectedDots) {
-            dotsState.brailleDots
-        }
+        val viewModelFactory =
+            InputViewModelFactory(
+                application,
+                expectedDots
+            ) {
+                dotsState.brailleDots
+            }
         viewModel = ViewModelProvider(
-            this@InputDotsFragment, viewModelFactory
+            this@InputSymbolFragment, viewModelFactory
         ).get(InputViewModel::class.java)
         buzzer = activity?.getSystemService()
 
 
         inputViewModel = viewModel
-        lifecycleOwner = this@InputDotsFragment
+        lifecycleOwner = this@InputSymbolFragment
 
 
         prevButton.setOnClickListener {
@@ -101,7 +109,7 @@ class InputDotsFragment : AbstractStepFragment(R.string.lessons_help_input_dots)
             dotsState
         ) {
             val notify = {
-                makeIncorrectToast()
+                makeIncorrectLetterToast(symbol.symbol.toString())
                 buzzer.checkedBuzz(preferenceRepository.incorrectBuzzPattern, preferenceRepository)
             }
             if (userTouchedDots) notify()
@@ -117,7 +125,9 @@ class InputDotsFragment : AbstractStepFragment(R.string.lessons_help_input_dots)
 
         viewModel.observeEventPassHint(
             viewLifecycleOwner, dotsState
-        )
+        ) {
+            makeIntroLetterToast(symbol.symbol.toString())
+        }
 
     }.root
 }
