+++ android/app/src/main/java/com/github/braillesystems/learnbraille/ui/screens/theory/steps/InputViewModel.kt
@@ -1,4 +1,4 @@
-package com.github.braillesystems.learnbraille.ui.screens.theory
+package com.github.braillesystems.learnbraille.ui.screens.theory.steps
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
@@ -18,7 +18,11 @@ class InputViewModelFactory(
     override fun <T : ViewModel> create(modelClass: Class<T>): T =
         if (modelClass.isAssignableFrom(InputViewModel::class.java)) {
             @Suppress("UNCHECKED_CAST")
-            InputViewModel(application, getEnteredDots, expectedDots) as T
+            InputViewModel(
+                application,
+                getEnteredDots,
+                expectedDots
+            ) as T
         } else {
             throw IllegalArgumentException("Unknown ViewModel class")
         }
