+++ android/app/src/main/res/navigation/navigation.xml
@@ -68,7 +68,7 @@
 
     <fragment
         android:id="@+id/infoFragment"
-        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.InfoFragment"
+        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.steps.InfoFragment"
         android:label="InfoFragment">
         <argument
             android:name="step"
@@ -76,7 +76,7 @@
     </fragment>
     <fragment
         android:id="@+id/inputDotsFragment"
-        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.InputDotsFragment"
+        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.steps.InputDotsFragment"
         android:label="InputDotsFragment">
         <argument
             android:name="step"
@@ -84,7 +84,7 @@
     </fragment>
     <fragment
         android:id="@+id/inputSymbolFragment"
-        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.InputSymbolFragment"
+        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.steps.InputSymbolFragment"
         android:label="InputSymbolFragment">
         <argument
             android:name="step"
@@ -92,7 +92,7 @@
     </fragment>
     <fragment
         android:id="@+id/showDotsFragment"
-        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.ShowDotsFragment"
+        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.steps.ShowDotsFragment"
         android:label="ShowDotsFragment">
         <argument
             android:name="step"
@@ -100,7 +100,7 @@
     </fragment>
     <fragment
         android:id="@+id/showSymbolFragment"
-        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.ShowSymbolFragment"
+        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.steps.ShowSymbolFragment"
         android:label="ShowSymbolFragment">
         <argument
             android:name="step"
@@ -108,7 +108,7 @@
     </fragment>
     <fragment
         android:id="@+id/lastInfoFragment"
-        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.LastInfoFragment"
+        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.steps.LastInfoFragment"
         android:label="LastInfoFragment">
         <argument
             android:name="step"
@@ -116,7 +116,7 @@
     </fragment>
     <fragment
         android:id="@+id/firstInfoFragment"
-        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.FirstInfoFragment"
+        android:name="com.github.braillesystems.learnbraille.ui.screens.theory.steps.FirstInfoFragment"
         android:label="FirstInfoFragment">
         <argument
             android:name="step"
